Fixed Issue 218 (http://code.google.com/p/beast-mcmc/issues/detail?id=218).

+++ trunk/src/dr/app/beast/BeastDialog.java
@@ -164,7 +164,7 @@ public class BeastDialog {
         beagleResourceCombo.setSelectedItem("GPU");
     }
 
-    public boolean showDialog(String title, int seed) {
+    public boolean showDialog(String title, long seed) {
 
         JOptionPane optionPane = new JOptionPane(optionPanel,
                 JOptionPane.PLAIN_MESSAGE,
